Refactor username-screen
